- alle htmx herschreven en vervangen door ajax op de catalogus pagina
- htmx  vervangen op categorieen door ajax

- event delegation toegevoegd om ajax call naar get page handler te centraliseren

- Weergave van ingestelde filters toegevoegd om gebruiker te laten zien waarop hij/zij momenteel filtert

-  nog wat zaken denk ik

+++ ProjectCore/Features/ProductAttributen/Kleuren/Commands/DeleteColors.cs
@@ -19,6 +19,7 @@ namespace ProjectCore.Features.ProductAttributen.Kleuren.Commands
             public CommandValidator()
             {
                 RuleFor(c => c.KleurenToDelete).NotNull().NotEmpty().WithMessage("Er werden geen kleuren geselecteerd om te verwijderen");
+                
             }
         }
 
@@ -40,6 +41,14 @@ namespace ProjectCore.Features.ProductAttributen.Kleuren.Commands
                     if (!validationResult.IsValid)
                     {
                         return new ValidationErrorResult("Validatie mislukt voor het verwijderen van kleuren", validationResult.Errors.Select(x => new ValidationError(x.PropertyName, x.ErrorMessage)).ToList());
+                    }
+                    //Niet toelaten dat standaardkleur verwijderd wordt
+                    var standaardKleur = await _context.ProductKleuren
+                                        .Where(k => k.Kleur.Equals("Standaard"))
+                                        .FirstOrDefaultAsync();
+                    if (request.KleurenToDelete.Contains(standaardKleur.Id))
+                    {
+                        return new ErrorResult("Standaardkleur mag niet worden verwijderd.");
                     }
                     var numberOfKleurenDeleted = await _context.ProductKleuren.Where(pk => request.KleurenToDelete.Contains(pk.Id)).ExecuteDeleteAsync();
                     return new SuccessResult<int>(numberOfKleurenDeleted);
